Add files via upload
Initial upload... polling of single port pin. 4 defines in header need to be set. Some people have reported timings out of spec for their devices, if issues persist move device close to MCU with pull-up resistor and adjust timings in Cpp file if necessary. Short description of how to use the library is in the header file.